# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RideTech API",
        "description": "Документация API для Инноватсияи ояндасоз",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/reviews/{driver_id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Получить список отзывов о водителе",
                "description": "Возвращает все отзывы о водителе.",
                "operationId": "34f483e778fdeccc9c342bf8c2d96eac",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "ID водителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список отзывов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Создать отзыв о водителе",
                "description": "Создаёт новый отзыв для водителя.",
                "operationId": "06522b96056c6c8047799fd8ac678b70",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "ID водителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Отзыв успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateReviewRequest": {
                "required": [
                    "trip_id",
                    "rating"
                ],
                "properties": {
                    "trip_id": {
                        "description": "ID поездки, к которой относится отзыв",
                        "type": "integer"
                    },
                    "rating": {
                        "description": "Оценка от 1 до 5",
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "comment": {
                        "description": "Комментарий к отзыву",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "trip_id": {
                        "type": "integer"
                    },
                    "driver_id": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "BearerAuth": []
        }
    ]
}